requirements.txt
accelerate==0.23.0
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
filelock==3.12.4
frozenlist==1.4.0
fsspec==2023.9.2
hiredis==2.2.3
huggingface-hub==0.17.3
idna==3.4
Jinja2==3.1.2
magic-filter==1.0.12
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.26.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.2.140
nvidia-nvtx-cu12==12.1.105
packaging==23.2
peft==0.5.0
prometheus-client==0.17.1
psutil==5.9.5
pydantic==2.3.0
pydantic_core==2.6.3
PyYAML==6.0.1
redis==5.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
sympy==1.12
tokenizers==0.14.1
torch==2.1.0
tqdm==4.66.1
transformers==4.34.1
triton==2.1.0
typing_extensions==4.7.1
urllib3==2.0.6
yarl==1.9.2
bitsandbytes==0.41.1
sentencepiece==0.1.99
scipy==1.11.3
protobuf==4.24.4
